requirements.txt
absl-py==1.0.0
alabaster==0.7.12
alembic==1.7.7
anykeystore==0.2
appdirs==1.4.4
attrs==22.1.0
audioread==3.0.0
autopage==0.5.0
autopep8==1.6.0
Babel==2.10.1
black==19.10b0
brotlipy==0.7.0
cachetools==4.2.4
category-encoders==2.4.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.5
colorlog==6.6.0
cryptography==37.0.1
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dgl==0.8.1
dglgo==0.0.1
docker-pycreds==0.4.0
docutils==0.17.1
einops==0.6.1
filelock==3.6.0
fonttools==4.32.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
huggingface-hub==0.5.1
hupper==1.10.3
idna==3.4
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.2
librosa==0.9.2
llvmlite==0.39.1
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mypy-extensions==0.4.3
networkx==2.8.4
numba==0.56.2
numpy==1.21.5
numpydoc==1.3.1
oauthlib==3.2.0
optuna==2.10.0
packaging==21.3
pandas==1.4.2
PasteDeploy==2.1.1
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.2
pbr==5.8.1
Pillow==9.1.0
pip==22.1.2
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
pooch==1.6.0
prettytable==3.2.0
promise==2.3
protobuf==3.20.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyramid==2.0
pyramid-mailer==0.15.1
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.9.13
repoze.sendmail==4.4.1
requests==2.28.1
requests-oauthlib==1.3.0
resampy==0.4.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.49
scikit-learn==1.1.2
scipy==1.8.0
seaborn==0.11.2
senticnet==1.6
sentry-sdk==1.5.8
setproctitle==1.2.2
setuptools==63.4.1
shortuuid==1.0.8
six==1.16.0
sklearn2==0.0.13
smmap==5.0.0
snowballstemmer==2.2.0
SoundFile==0.10.3.post1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.32
statsmodels==0.12.2
stevedore==3.5.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
timm==0.5.4
tokenizers==0.12.0
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.10.0
tornado==6.1
tqdm==4.64.1
transaction==3.0.1
transformers==4.17.0
translationstring==1.4
typed-ast==1.5.4
typer==0.4.1
typing_extensions==4.3.0
urllib3==1.26.11
velruse==1.1.1
venusian==3.0.0
visdom==0.1.8.9
wandb==0.12.11
wcwidth==0.2.5
WebOb==1.8.7
websocket-client==1.3.1
Werkzeug==2.0.1
wheel==0.37.1
word2vec==0.11.1
wordninja==2.0.0
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xlrd==2.0.1
xlrd2==1.3.4
xlwt==1.3.0
yapf==0.32.0
yaspin==2.1.0
yellowbrick==1.5
zipp==3.8.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
